+++ bom/application/pom.xml
@@ -615,11 +615,6 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-arc-kotlin</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>io.quarkus</groupId>
                 <artifactId>quarkus-arc-dev</artifactId>
                 <version>${project.version}</version>
             </dependency>
