+++ extensions/kotlin/deployment/src/test/kotlin/io/quarkus/kotlin/arc/RequestContextCoroutineContextTest.kt
@@ -1,4 +1,4 @@
-package io.quarkus.arc.kotlin
+package io.quarkus.kotlin.arc
 
 import io.quarkus.arc.Arc
 import io.quarkus.test.QuarkusUnitTest
